prerender: kill errored worker, express-dom might need a pool of two

+++ packages/prerender/index.js
@@ -1,5 +1,5 @@
 const Path = require('path');
-const { Pool, TimeoutError } = require('tarn');
+const { Pool } = require('tarn');
 const fork = require('child_process').fork;
 
 module.exports = function(opt) {
@@ -29,6 +29,9 @@ module.exports = function(opt) {
 	const workerPath = Path.join(__dirname, 'worker.js');
 
 	const pool = new Pool({
+		validate: function(child) {
+			return !child.killed;
+		},
 		create: function(cb) {
 			var child;
 			try {
@@ -47,8 +50,9 @@ module.exports = function(opt) {
 			cb(null, child);
 		},
 		destroy: function(child) {
-			child.kill();
+			if (!child.killed) child.kill();
 		},
+		acquireTimeoutMillis: 5000,
 		min: 2,
 		max: 8
 	});
@@ -77,6 +81,7 @@ module.exports = function(opt) {
 			});
 			worker.once("error", function(err) {
 				worker.removeAllListeners("message");
+				worker.kill();
 				pool.release(worker);
 				next(err);
 			});
